pyproject.toml
[project]
name = "excelprocessor"
version = "0.1.0"
description = "Excel processor I made for myself "
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "actualpy>=0.17.0",
    "openpyxl==3.1.5",
    "pandas==2.2.3",
    "python-dotenv>=1.2.1",
    "pypdf>=5.0.0",
    "faker>=37.12.0",
]

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "basedpyright>=1.37.2",
    "flake8>=7.3.0",
    "mypy>=1.19.1",
    "pandas-stubs>=2.2.2.240807",
    "pylint>=3.3.9",
    "pyright>=1.1.408",
    "pytest>=8.4.2",
    "types-openpyxl>=3.1.5.20250919",
    "vulture>=2.14",
]
